1st Semester/Advanced Methods in Data Analysis/Project1/requirements.txt
torch~=2.5.1
torchvision~=0.20.1
opencv-python~=4.10.0.84
pandas~=2.2.3
ultralytics~=8.3.38
jupyter~=1.1.1
pillow~=11.0.0
matplotlib~=3.9.2
wandb~=0.18.7
PyYAML~=6.0.2